# [Pytoch][Vulkan] Create context for conv1d