--- a/aten/src/ATen/native/vulkan/ops/Register.h
+++ b/aten/src/ATen/native/vulkan/ops/Register.h
@@ -6,6 +6,7 @@ namespace vulkan {
 namespace ops {
 
 int register_vulkan_conv2d_packed_context();
+int register_vulkan_conv1d_packed_context();
 int register_vulkan_linear_packed_context();
 int register_vulkan_layernorm_packed_context();